namespace and version controllers and endpoints; use constraints
