+++ app/controllers/api/attendance_records_controller.rb
@@ -1,0 +1,6 @@
+module Api
+  module V1
+    class AttendanceRecordsController < ApplicationController
+    end
+  end
+end
