+++ app/controllers/api/cohorts_controller.rb
@@ -1,0 +1,14 @@
+module Api  
+  module V1
+    class CohortsController < ApplicationController
+
+      def index
+       render json: Cohort.all
+      end
+
+      def show
+        render json: Cohort.find(params[:id])
+      end
+    end
+  end
+end
