+++ app/controllers/api/holidays_controller.rb
@@ -1,0 +1,7 @@
+module Api
+  module V1
+
+    class HolidaysController < ApplicationController
+    end
+  end
+end
