+++ app/controllers/api/schedules_controller.rb
@@ -1,0 +1,7 @@
+module Api
+  module V1
+    class SchedulesController < ApplicationController
+    end
+  end
+end
+
