+++ app/controllers/api/users_controller.rb
@@ -1,0 +1,6 @@
+module Api
+  module V1
+    class UsersController < ApplicationController
+    end
+  end
+end
