--- app/controllers/attendance_records_controller.rb
@@ -1,2 +1,0 @@
-class AttendanceRecordsController < ApplicationController
-end
