--- app/controllers/cohorts_controller.rb
@@ -1,10 +1,0 @@
-class CohortsController < ApplicationController
-
-  def index
-   render json: Cohort.all
-  end
-
-  def show
-    render json: Cohort.find(params[:id])
-  end
-end
