--- app/controllers/holidays_controller.rb
@@ -1,2 +1,0 @@
-class HolidaysController < ApplicationController
-end
