+++ config/routes.rb
@@ -1,11 +1,16 @@
+require 'api_constraints'
 Rails.application.routes.draw do
-  resources :holidays
-  resources :attendance_records
-  resources :schedules
-  resources :cohorts
-  resources :users
+
+  namespace :api do
+    scope module: :v1, constraints: ApiConstraints.new(version: 1, default: true) do
+      resources :holidays
+      resources :attendance_records
+      resources :schedules
+      resources :cohorts
+      resources :users
+    end
+  end
 
   get '/login', to: "sessions#new", as: "login"
   get '/auth/:provider/callback', to: "sessions#create"
-  # For details on the DSL available within this file, see http://guides.rubyonrails.org/routing.html
 end
